+++ EDAcoin/Node.cpp
@@ -5,7 +5,7 @@ Node::dettachConnection(Node* connection)
 {
 	bool done = false;
 	list<Node*>::iterator itr = connections.begin();
-	int size = connections.size();
+	int size = (int) connections.size();
 	for (int i = 0; i < size && !done; i++, itr++)
 	{
 		if (*itr == connection)
